[proliferate] consolidate reviews MCP server
